ccnl-pkt: make centrality TLV 16-bit

+++ src/ccnl-core/include/ccnl-pkt.h
@@ -84,7 +84,7 @@ struct ccnl_pktdetail_ndntlv_s {
     /* Data */
     uint64_t freshnessperiod;      /**< defines how long a node has to wait (after the arrival of this data before) marking it “non-fresh” */
 #ifdef CACHING_ABC
-    uint8_t centrality;
+    uint16_t centrality;
 #endif
 };
